giza,karnak,mustang: add experimental files for them - not working
giza and karnak are based on the douglas files and mustang on the
mustang files from the upstream tree. the two versions for karnak
are for the v3.18 and the v4.9 kernel and the two for mustang are
for a 32bit and a 64bit kernel. the initial kernel configs are
from a simple make system_defconfig of the corresponding tree.

all this is just ment as the base for a playground to play around
with those and will most probably not work or even compile in this
state yet ...